does splitting it up help? (passes locally, but github actions fails)
